pyproject.toml
[project]
name = "dq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.22.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.1",
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.6",
    "types-pyyaml>=6.0.12.20241230",
]


[tool.mypy]
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true


[project.scripts]
dq = "mdp.cli.cli:main"
